actuator: make master resource optional
Actuator's master resource is now optional. It should have been optional
from the start, but code generation for optional object links did not
work correctly.

This change does not really need an ObjectVersion change in my opinion.
The IC24 eco is the only device which makes use of it, and it will work
with both definitions on the server just fine.