definitions: Add standard LwM2M objects to the registry
We want to distribute the OMA standard LwM2M objects in this repository
to allow gateway independent updates in the future. The standard LwM2M
objects are still also distributed with the Wakaama Python package in the
sg-bnw-lwm2mserver repository for testing and development purposes.

These files are a direct copy from the OMA LightweightM2M (LwM2M) Object
and Resource Registry.
https://technical.openmobilealliance.org/OMNA/LwM2M/LwM2MRegistry.html